Updated DataRefs/Commands loading when data is changed ...
There can be situation when the text files with data are changed outside the plug-in
for example you have edit it manually with a text editor and  after it
the plug-in doesn't have those new changes loaded.
So when the plug-in tries to save some changes it tries to load new data from the files before.