Added new abstraction for winapi ...
It is another attempt to make easer working with the GUI.
Unfortunately until we stop supporting 3Ds Max versions less than 2018 we need to support winapi instead of QT.
This new abstraction hides working with controls and windows procedures and it may be uncompleted and it assumes that necessary parts will be written when they are needed.
It doesn't support all necessary use case yet and has some untested parts.